Checkpoint 23/11

+++ Enplace/Enplace/Enplace.Service/Entities/UserMenuRecipes.cs
@@ -4,7 +4,7 @@
     {
         public int MenuID { get; set; }
         public int RecipeID { get; set; }
-        public int Scale { get; set; } = 1;
+        public decimal Scale { get; set; } = 1;
         public virtual UserMenu? Menu { get; set; }
         public virtual Recipe? Recipe { get; set; }
     }
